Expose problemStr
Useful for client code that has a js.spec problem and want to
convert it to string, having no need for using explainStr directly.
This patch also adds a couple of test for making sure that we don't have
undefined or [object Object] when the spec does not validate (should be more
thorough though).